requirements.txt
alembic==1.8.1
aniso8601==9.0.1
APScheduler==3.10.4
attrs==22.1.0
Authlib==1.3.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
cx-Oracle==8.3.0
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.1.3
Flask==2.0.2
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-oidc==1.4.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
htmlmin==0.1.12
httplib2==0.21.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.0.0
Jinja2==3.0.2
jsmin==3.0.1
jsonschema==4.9.1
lesscpy==0.15.0
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.2
oauth2client==4.1.3
pandas==2.2.0
pandasql==0.7.3
paramiko==2.11.0
ply==3.11
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-keycloak==2.6.0
pytz==2022.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.29
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.11
Werkzeug==2.0.3
WTForms==3.0.0
xxhash==3.0.0
zipp==3.8.1